Don't expose all wrapped functions, and define them in the current namespace
Wrapping all lib functions in namespace crpcut::wrapped was a major mistake.
It places the burden of the test case writer to know which functions are
wrapped internally by crpcut. By moving them to a namespace controlled by
the test case writer, there will be no collisons (and it's not an error if
a test wraps a function that crpcut already wraps.)