Fixed bugs when a test forked a naughty child
More specifically, if the child of a test called for crpcut's attention,
for example by failing an ASSERT_EQ, crpcut main process got very confused
and crashed.

If the child of a test went into an infinite loop, it would make the
test case process a zombie if it was killed by crpcut due to timeout.
Since zombies don't die easily, crpcut never finished.